pyproject.toml
[project]
name = "imagedownload"
version = "0.1.0"
description = "Google Image Search and Download Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp-server>=0.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
